+++ include/ss_main.h
@@ -424,7 +424,7 @@ void sscrnModelClear(SUB_SCREEN* wk);
 void sscrnLightClear(SUB_SCREEN* wk);
 void sscrnLightCreate(SUB_SCREEN* wk, cLit* lit);
 void sscrnMainMenuInit(SUB_SCREEN* wk, int no);
-void numDisp(u8 id, int num, Vec* pos, u32 flags);
+void numDisp(int id, int num, Vec* pos, u32 flags);
 void sscrnCameraInit(SUB_SCREEN* wk, Camera* cam);
 void sscrnModelFree(SUB_SCREEN* wk);
 void generalModelAlloc(SUB_SCREEN* wk);
